763 - Adds new model type for admin comments
After a discussion with the grad school, it became clear that a separate model
would be better than repurposing FormatReviewFile and FinalSubmissionFile. For one,
it would be safer, as we really don't want these files to show up on Explore. For
another, it would make presenting easier, as these files should be clearly distinct
from student submission.