857 - Tells Autoloaders in application.rb to ignore certain dirs
There is an issue with Zeitwerk where autoloaders expect the dirs and
files in lib to be organized a certain way, and it causes the application
to crash is eager_load is true. Setting Rails.autoloaders.main.ignore for these
seems to fix the issue. They are included in rails_helper.rb