Bugfix - moves extension methods to initializer folders
In the past, we had extension methods in the lib/ folder. However,
Zeitwerk expects a specific format based on the directory and file names
within the lib folder and was throwing errors. By moving the extension methods
to the initializer folder, we still have access to the extension methods without
encountering that error.