936 - Always show default final submission note
There was a check to make sure the default final submission note didn't override
a blank note if the submission was past a certain point. Unfortunately that check
was set to look if final submission was submitted (which would always be true),
so this corrects it so the check is now for whether the submission is beyond
waiting for final response.